+++ src/main/java/com/doublew2w/sbs/mybatis/executor/resultset/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 结果集：包含结果处理逻辑
+ */
+package com.doublew2w.sbs.mybatis.executor.resultset;