+++ src/main/java/com/doublew2w/sbs/mybatis/session/ResultHandler.java
@@ -1,0 +1,13 @@
+package com.doublew2w.sbs.mybatis.session;
+
+/**
+ * 结果处理器
+ *
+ * @author: DoubleW2w
+ * @date: 2024/9/4 3:54
+ * @project: sbs-mybatis
+ */
+public interface ResultHandler {
+  /** 处理结果 */
+  void handleResult();
+}