+++ src/main/java/com/doublew2w/sbs/mybatis/session/defaults/DefaultSqlSession.java
@@ -1,10 +1,13 @@
 package com.doublew2w.sbs.mybatis.session.defaults;
 
+import com.doublew2w.sbs.mybatis.executor.Executor;
 import com.doublew2w.sbs.mybatis.mapping.BoundSql;
 import com.doublew2w.sbs.mybatis.mapping.Environment;
 import com.doublew2w.sbs.mybatis.mapping.MappedStatement;
 import com.doublew2w.sbs.mybatis.session.Configuration;
 import com.doublew2w.sbs.mybatis.session.SqlSession;
+import lombok.extern.slf4j.Slf4j;
+
 import java.lang.reflect.Method;
 import java.sql.*;
 import java.util.ArrayList;
@@ -15,11 +18,14 @@ import java.util.List;
  * @date: 2024/9/1 5:41
  * @project: sbs-mybatis
  */
+@Slf4j
 public class DefaultSqlSession implements SqlSession {
-  private final Configuration configuration;
+  private Configuration configuration;
+  private Executor executor;
 
-  public DefaultSqlSession(Configuration configuration) {
+  public DefaultSqlSession(Configuration configuration, Executor executor) {
     this.configuration = configuration;
+    this.executor = executor;
   }
 
   @Override
@@ -29,23 +35,10 @@ public class DefaultSqlSession implements SqlSession {
 
   @Override
   public <T> T selectOne(String statement, Object parameter) {
-    try {
-      MappedStatement mappedStatement = configuration.getMappedStatement(statement);
-      Environment environment = configuration.getEnvironment();
-
-      Connection connection = environment.getDataSource().getConnection();
-
-      BoundSql boundSql = mappedStatement.getBoundSql();
-      PreparedStatement preparedStatement = connection.prepareStatement(boundSql.getSql());
-      preparedStatement.setLong(1, Long.parseLong(((Object[]) parameter)[0].toString()));
-      ResultSet resultSet = preparedStatement.executeQuery();
-
-      List<T> objList = resultSet2Obj(resultSet, Class.forName(boundSql.getResultType()));
-      return objList.get(0);
-    } catch (Exception e) {
-      e.printStackTrace();
-      return null;
-    }
+    log.info("数据库操作开始");
+    MappedStatement ms = configuration.getMappedStatement(statement);
+    List<T> list = executor.query(ms, parameter, Executor.NO_RESULT_HANDLER, ms.getBoundSql());
+    return list.get(0);
   }
 
   @Override
@@ -56,34 +49,5 @@ public class DefaultSqlSession implements SqlSession {
   @Override
   public Configuration getConfiguration() {
     return configuration;
-  }
-
-  private <T> List<T> resultSet2Obj(ResultSet resultSet, Class<?> clazz) {
-    List<T> list = new ArrayList<>();
-    try {
-      ResultSetMetaData metaData = resultSet.getMetaData();
-      int columnCount = metaData.getColumnCount();
-      // 每次遍历行值
-      while (resultSet.next()) {
-        T obj = (T) clazz.getDeclaredConstructor().newInstance();
-        for (int i = 1; i <= columnCount; i++) {
-          Object value = resultSet.getObject(i);
-          String columnName = metaData.getColumnName(i);
-          String setMethod =
-              "set" + columnName.substring(0, 1).toUpperCase() + columnName.substring(1);
-          Method method;
-          if (value instanceof Timestamp) {
-            method = clazz.getMethod(setMethod, Date.class);
-          } else {
-            method = clazz.getMethod(setMethod, value.getClass());
-          }
-          method.invoke(obj, value);
-        }
-        list.add(obj);
-      }
-    } catch (Exception e) {
-      e.printStackTrace();
-    }
-    return list;
   }
 }