+++ src/main/java/com/doublew2w/sbs/mybatis/session/defaults/DefaultSqlSessionFactory.java
@@ -1,9 +1,16 @@
 package com.doublew2w.sbs.mybatis.session.defaults;
 
 import com.doublew2w.sbs.mybatis.binding.MapperRegistry;
+import com.doublew2w.sbs.mybatis.executor.Executor;
+import com.doublew2w.sbs.mybatis.mapping.Environment;
 import com.doublew2w.sbs.mybatis.session.Configuration;
 import com.doublew2w.sbs.mybatis.session.SqlSession;
 import com.doublew2w.sbs.mybatis.session.SqlSessionFactory;
+import com.doublew2w.sbs.mybatis.session.TransactionIsolationLevel;
+import com.doublew2w.sbs.mybatis.transaction.Transaction;
+import com.doublew2w.sbs.mybatis.transaction.TransactionFactory;
+
+import java.sql.SQLException;
 
 /**
  * @author: DoubleW2w
@@ -20,6 +27,9 @@ public class DefaultSqlSessionFactory implements SqlSessionFactory {
 
   @Override
   public SqlSession openSession() {
-    return new DefaultSqlSession(configuration);
+    // 创建执行器
+    final Executor executor = configuration.newExecutor();
+    // 创建DefaultSqlSession
+    return new DefaultSqlSession(configuration, executor);
   }
 }